requirements.txt
attrdict==2.0.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
datasets==1.11.0
dill==0.3.4
Django==2.2.5
djangorestframework==3.10.3
filelock==3.0.12
fsspec==2021.7.0
fugashi==1.1.1
huggingface-hub==0.0.12
idna==3.2
ipadic==1.0.0
joblib==1.0.1
mojimoji==0.0.11
multiprocess==0.70.12.2
nlp==0.4.0
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pyarrow==5.0.0
pyknp==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.21
requests==2.26.0
sacremoses==0.0.45
six==1.16.0
sqlparse==0.4.1
tokenizers==0.10.3
torch==1.9.0
torchtext==0.10.0
tqdm==4.62.2
transformers==4.9.2
typing-extensions==3.10.0.0
urllib3==1.26.6
xxhash==2.0.2
